Minor test changes
