+++ crawlerapp/frontend/tests/unit/adapters/application-test.js
@@ -5,7 +5,7 @@ module('Unit | Adapter | application', function(hooks) {
   setupTest(hooks);
 
   // Replace this with your real tests.
-  test('it exists', function(assert) {
+  test('application adapter exists', function(assert) {
     let adapter = this.owner.lookup('adapter:application');
     assert.ok(adapter);
   });
