+++ crawlerapp/public/assets/frontend-f085bae5db76b5521f3dffc177eca6ab.css
@@ -1,0 +1,14 @@
+@import 'components/user-input-form-c6782d90c46df6a21e410e06e2f70e48.css';
+@import 'components/nav-bar-f4cde15da5f63207be826fbfb8ac51f6.css';
+
+.lower-components {
+    display: flex;
+}
+
+.history-component {
+    flex-basis:  40%;
+}
+
+.backend-list {
+    margin-left: 300px; 
+}
