+++ crawlerapp/public/index.html
@@ -8,7 +8,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
 
     
-<meta name="frontend/config/environment" content="%7B%22modulePrefix%22%3A%22frontend%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22/%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%7D%2C%22APP%22%3A%7B%22name%22%3A%22frontend%22%2C%22version%22%3A%220.0.0+4ee8b2c5%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
+<meta name="frontend/config/environment" content="%7B%22modulePrefix%22%3A%22frontend%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22%2F%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%7D%2C%22APP%22%3A%7B%22name%22%3A%22frontend%22%2C%22version%22%3A%220.0.0%2B3829f6c3%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
     <!-- Compiled and minified CSS -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
     
@@ -18,15 +18,15 @@
 
     <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
     <link integrity="" rel="stylesheet" href="/assets/vendor-d41d8cd98f00b204e9800998ecf8427e.css">
-    <link integrity="" rel="stylesheet" href="/assets/frontend-afd6b727b11e13bc767f0ebd17c47ecd.css">
+    <link integrity="" rel="stylesheet" href="/assets/frontend-f085bae5db76b5521f3dffc177eca6ab.css">
 
     
   </head>
   <body>
     
 
-    <script src="/assets/vendor-5bec10d902faf06565f5cc7b5fd59831.js" integrity="sha256-TnpXdN91gQjexjowM0kWvDtLVl/pE3RaeiiTG+CpuDE= sha512-D3Lu1Kq95RQkc3wUx1beq0wY/SSJ8JMCAvMIViJDH4pQ5r+b+d3V/ewUw8Vaev1htyoHf9fn0Us4K1S0OTk00g==" ></script>
-    <script src="/assets/frontend-c31ff617789324c6e42eef934420851b.js" integrity="sha256-Gs/Jb76DHEpH3xDG3pIa6salIUlAygvx3hfANnSJP/Y= sha512-/Nj96smCvtp6XhcfKDx42CHRjF+n49TezkLj2toEXIKT9uibYeESw+Og+rbNDgnAV9rfa3wI6xqBiozlHzIuUQ==" ></script>
+    <script src="/assets/vendor-aebbd2cbe6b1526a2c6895c40a0fb6f3.js" integrity="sha256-9GvXDAk57yg+p/iUYUqUucpNLq8krVDQNISNU79sHWo= sha512-NLDwInvPkzssdy98XHroO0LAhQbwqbCdvEqO69wkIe2OMgABiROJbfC6CwoZUttqLRBsDmtUXfxib0ErVVeHZw==" ></script>
+    <script src="/assets/frontend-5d6976a1caae79b6821fef1d6345acbc.js" integrity="sha256-DZ6KpGWU9PV+w//+j6qRUzxX/wRqF8qzTutXVqH17RY= sha512-ICCEStRrsrB/sJU2mHuzsHNUUEn4V19kJH2tD6CjM+YN4fP64rV5k5l+A8UNI4JP3j8o2qXv0m90q7aka3kCAg==" ></script>
 
     
   </body>
